+++ test/jest/helpers/index.js
@@ -1,3 +1,4 @@
 export * from './renderWithReduxForm';
+export * from './renderWithFinalForm';
 export * from './stripesMock';
 export * from './translationsProperties';